requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
pydantic==2.5.0
pydantic-settings==2.1.0
python-multipart==0.0.6
jinja2==3.1.2

# Data processing
pandas==2.1.3
numpy==1.26.2
openpyxl==3.1.2

# Machine Learning
scikit-learn==1.3.2
xgboost==2.0.2
lightgbm==4.1.0
joblib==1.3.2

# Deep Learning / NLP
torch==2.1.1
transformers==4.35.2
sentencepiece==0.1.99

# Vector Search
faiss-cpu==1.7.4

# LLM API
groq==0.4.1

# Utilities
python-dotenv==1.0.0